finished opcodes
